Add stage-3 not included in env
