+++ services/storage/meta-index/package.json
@@ -11,11 +11,12 @@
     "babel-core": "6.26.0",
     "babel-loader": "7.1.2",
     "babel-plugin-transform-runtime": "6.23.0",
-    "babel-preset-es2015": "6.24.1",
+    "babel-preset-env": "1.6.0",
     "babel-preset-stage-3": "6.24.1",
     "jest": "21.2.1",
     "serverless": "1.23.0",
     "serverless-dynamodb-autoscaling": "0.6.2",
+    "serverless-jest-plugin": "0.1.6",
     "serverless-webpack": "3.1.1",
     "webpack": "3.6.0"
   },
