+++ services/storage/signed-uploads/.babelrc
@@ -1,4 +1,4 @@
 {
   "plugins": ["transform-runtime"],
-  "presets": ["es2015", "stage-3"]
+  "presets": ["env", "stage-3"]
 }
