+++ services/storage/signed-uploads/package.json
@@ -1,20 +1,23 @@
 {
   "private": true,
-  "name": "@eubfr/signed-uploads",
+  "name": "@eubfr/storage-signed-uploads",
   "description": "",
   "version": "0.0.1",
   "scripts": {
     "deploy": "sls deploy -v",
-    "start": "sls offline start"
+    "start": "sls offline start",
+    "test:unit": "jest --testPathPattern=unit"
   },
   "devDependencies": {
     "aws-sdk": "2.127.0",
     "babel-core": "6.26.0",
     "babel-loader": "7.1.2",
     "babel-plugin-transform-runtime": "6.23.0",
-    "babel-preset-es2015": "6.24.1",
+    "babel-preset-env": "1.6.0",
     "babel-preset-stage-3": "6.24.1",
+    "jest": "21.2.1",
     "serverless": "1.23.0",
+    "serverless-jest-plugin": "0.1.6",
     "serverless-offline": "3.16.0",
     "serverless-webpack": "3.1.1",
     "webpack": "3.6.0"
