--- services/storage/signed-uploads/src/docroot.js
@@ -1,16 +1,0 @@
-/* eslint-disable import/prefer-default-export */
-export const handler = (event, context, cb) => {
-  const response = {
-    statusCode: 200,
-    headers: {
-      'Access-Control-Allow-Origin': '*', // Required for CORS support
-      'Access-Control-Allow-Credentials': true, // Required for cookies, authorization headers with HTTPS
-    },
-    body: JSON.stringify({
-      message:
-        'Docroot is not used at the moment. Please check the other endpoints',
-    }),
-  };
-
-  cb(null, response);
-};
